factor SilentAuthenticationCredential out of SharedTokenCacheCredential
